Liq migration + refactoring + some code  of Restaurant Controller

+++ src/main/java/ru/javaops/topjava/HasId.java
@@ -4,9 +4,9 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import org.springframework.util.Assert;
 
 public interface HasId {
-    Integer getId();
+    Long getId();
 
-    void setId(Integer id);
+    void setId(Long id);
 
     @JsonIgnore
     default boolean isNew() {
@@ -14,7 +14,7 @@ public interface HasId {
     }
 
     // doesn't work for hibernate lazy proxy
-    default int id() {
+    default Long id() {
         Assert.notNull(getId(), "Entity must has id");
         return getId();
     }
